requirements.txt
build
twine
python-decouple
termcolor